+++ docs/changes.rst
@@ -8,6 +8,7 @@ Weblate 5.14
 * :ref:`addon-weblate.flags.target_repo_update` is now available as an add-on.
 * :http:delete:`/api/projects/(string:project)/labels/(int:label_id)/`.
 * Added multivalue CSV in :ref:`download` for multivalue formats.
+* :ref:`file_format_params` can now be configured via the :ref:`api`.
 
 .. rubric:: Improvements